Feature/fwpm 1622 add adapter backlog limits (#87)
Add adapter for BacklogLimits whit searchEntities client method

+++ src/main/java/com/mercadolibre/flow/control/tool/feature/backlog/monitor/dto/ProcessLimit.java
@@ -4,8 +4,8 @@ import com.mercadolibre.flow.control.tool.feature.entity.ProcessName;
 
 public record ProcessLimit(
     ProcessName name,
-    int lower,
-    int upper
+    Long lower,
+    Long upper
 ) {
 
 }
